Add import test that does not import addresses
git-svn-id: http://svn.code.sf.net/p/conferenceplan/code/PlannerPrototype@77 1f962665-22ff-41fc-9765-37673d54b4f3

+++ app/controllers/pending_import_people_controller.rb
@@ -47,8 +47,21 @@ class PendingImportPeopleController < ApplicationController
      @parsed_file=CSV::Reader.parse(params[:dump][:file])
      n=0
      @parsed_file.shift
+     @testImport = false
+     if (params[:dump][:importtype] == "renotest")
+       @testImport = true
+     elsif (params[:dump][:importtype] != "reno")
+        redirect_to :action => 'index'
+     end
+     
      @parsed_file.each  do |row|  
-       pendingImportPerson = PendingImportPerson.new(:first_name => row[2],
+       if (@testImport)
+          pendingImportPerson = PendingImportPerson.new(:first_name => row[2],
+                                                     :last_name => row[3],
+                                                     :registration_number => row[0],
+                                                     :registration_type => row[1])
+       else
+          pendingImportPerson = PendingImportPerson.new(:first_name => row[2],
                                                      :last_name => row[3],
                                                      :line1 => row[4],
                                                      :line2 => row[5],
@@ -60,20 +73,9 @@ class PendingImportPeopleController < ApplicationController
                                                      :registration_number => row[0],
                                                      :registration_type => row[1],
                                                      :email => row[11])
+       end
         pendingImportPerson.save
- #      pendingImportPerson.first_name = row[2]
- #      pendingImportPerson.last_name = row[3]
- #      pendingImportPerson.line1 = row[4]
- #      pendingImportPerson.line2 = row[5]
- #      pendingImportPerson.city  = row[6]
- #      pendingImportPerson.postcode = row[7]
- #      pendingImportPerson.state = row[8]
- #      pendingImportPerson.country = row[9]
- #      pendingImportPerson.phone = row[10]
- #      pendingImportPerson.registration_number = row[0]
- #      pendingImportPerson.registration_type = row[1]
- #      pendingImportPerson.email = row[11]
-  
+
        newPersonFlag = true
        savePersonAsPending = false
        nReg = nil
@@ -81,45 +83,45 @@ class PendingImportPeopleController < ApplicationController
           nReg = RegistrationDetail.find_or_initialize_by_registration_number(pendingImportPerson.registration_number)
           newPersonFlag = nReg.new_record?
        end
-#       pendingImportPerson.save()
       
        if (newPersonFlag)
          # search for people with same name
          @matchpeople = Person.find_all_by_last_name_and_first_name(pendingImportPerson.last_name,pendingImportPerson.last_name)
+
          if (@matchpeople.length != 0)
-             # check if address is the same
-             @matchpeople.each do |matchperson|
-                @matchperson.postal_addresses.each do |matchaddress|
-                  if ((pendingImportPerson.line1 == matchaddress.line1) &&
-                      (pendingImportPerson.line2 == matchaddress.line2) &&
-                      (pendingImportPerson.line3 == matchaddress.line3) &&
-                       (pendingImportPerson.city == matchaddress.city) &&
-                       (pendingImportPerson.state == matchaddress.state) &&
-                       (pendingImportPerson.postcode == matchaddress.postcode) &&
-                       (pendingImportPerson.country == matchaddress.country)) 
-                      newPersonFlag = false
-                  end
+             if (pendingImportPerson.addressNil? && pendingImportPerson.emailNil? && @testImport == true)
+               newPersonFlag == false
+             else               
+                # check if address is the same
+                @matchpeople.each do |matchperson|
+                   @matchperson.postal_addresses.each do |matchaddress|
+                     if (pendingImportPerson.addressMatch?(matchaddres))
+                         newPersonFlag = false
+                     end
+                   end
                 end
-            end
-            if (newPersonFlag)
-              @matchpeople.each do |matchperson|
-                @matchperson.email_addresses.each do |matchaddress|
-                  if (pendingImportPerson.email == matchaddress.email)
-                    newPersonFlag = false
-                  end #if
-                end #@matchaddress
-              end #@matchpeople
-            end #newPeopleFlag
-            
-            # we found a name match, but the address did not match,
-            # so we should just save it as pending
-            if (newPersonFlag)
-              savePersonAsPending = true
-              newPersonFlag = false
+             
+                if (newPersonFlag)
+                  @matchpeople.each do |matchperson|
+                     @matchperson.email_addresses.each do |matchaddress|
+                       if (pendingImportPerson.email == matchaddress.email)
+                         newPersonFlag = false
+                       end #if
+                     end #@matchaddress
+                  end #@matchpeople
+                end #newPeopleFlag
+              
+                # we found a name match, but the address did not match,
+                # so we should just save it as pending
+               if (newPersonFlag)
+                  savePersonAsPending = true
+                  newPersonFlag = false
+               end
             end
          end
        end
        
+ 
        if (newPersonFlag)
            p=Person.new(:first_name => pendingImportPerson.first_name,
                         :last_name => pendingImportPerson.last_name)
@@ -129,20 +131,22 @@ class PendingImportPeopleController < ApplicationController
            end
            
            p.save
-           @a=p.postal_addresses.new(:line1 => pendingImportPerson.line1,
-                                     :line2 => pendingImportPerson.line2,
-                                     :city  => pendingImportPerson.city,
-                                     :postcode => pendingImportPerson.postcode,
-                                     :state => pendingImportPerson.state,
-                                     :country => pendingImportPerson.country,
-                                     :phone => pendingImportPerson.phone)    
-           @e=p.email_addresses.new(:email => pendingImportPerson.email)  
-           if (p.save)
-              pendingImportPerson.destroy
-              n=n+1
-              GC.start if n%50==0
-           end
-           flash.now[:message]="CSV Import Successful,  #{n} new
+           if (@testImport == false)
+              @a=p.postal_addresses.new(:line1 => pendingImportPerson.line1,
+                                        :line2 => pendingImportPerson.line2,
+                                        :city  => pendingImportPerson.city,
+                                        :postcode => pendingImportPerson.postcode,
+                                        :state => pendingImportPerson.state,
+                                        :country => pendingImportPerson.country,
+                                        :phone => pendingImportPerson.phone)    
+              @e=p.email_addresses.new(:email => pendingImportPerson.email)
+            end
+            if (p.save)
+                 pendingImportPerson.destroy
+                 n=n+1
+                 GC.start if n%50==0
+            end
+            flash.now[:message]="CSV Import Successful,  #{n} new
                                 records added to data base"
        else
          if (!savePersonAsPending)
@@ -152,15 +156,19 @@ class PendingImportPeopleController < ApplicationController
             # save pending person for later updating
             person = Person.find_by_id_and_last_name_and_first_name(nReg.person_id,pendingImportPerson.last_name,pendingImportPerson.first_name)
             if (person)
-              if (person.UpdateIfPendingPersonDifferent(pendingImportPerson.id))
+              if (@testImport == true)
                 pendingImportPerson.destroy
-              end
-            end
-           else
-              pendingImportPerson.destroy
-           end
-       end
-      end
+              else
+                if (person.UpdateIfPendingPersonDifferent(pendingImportPerson.id))
+                  pendingImportPerson.destroy
+                end #update
+              end #testimport
+             end #person (we did not find a name and id match, so we probably have a bad id)
+         else 
+           pendingImportPerson.destroy
+         end #reg nil
+       end #save person as pending
+      end #newperson
      end #do
      redirect_to :action => 'index'
  end
